requirements.txt
arxiv
requests
sentence-transformers
gradio
google-generativeai
Pillow
openai